DOC: Fix bug in maintenance tool update documentation
The extraction of maintenance tool to a temporary folder is required
only for macOs as Windows already has deferred rename and Linux uses
inode. Fixed the bug in documetion where the maintenance tool extract to
a temporary folder was performed in all platforms.

Also fixed the update.rcc location in macOs.

Task-number: QTIFW-3715
Change-Id: I5e7054dd8833be119af856be1f6b896809f5fe15
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>