Add size information to component summary view
Size information is shown in component summary view cumulatively, as
there can be virtual items under visual items, the size does not match
to the total value unless calculated cumulatively.
Also fixed a bug where uninstallable component is not shown in summary
view if it is not a directly unselected component.

Task-number: QTIFW-3752
Change-Id: I2090e5abfd13faea05e4df9480503432cc92d01a
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>