CLI: Fix unnecessary uninstall of virtual, non-tristate components
Virtual item was incorrectly set as unchecked item, although it was
installed. This caused the item to be uninstalled when other items were
uninstalled. As the virtual component was selected for uninstallation,
uninstallcalculator noticed that dependency component is uninstalled,
and caused a chain reaction so also other components which were
dependent on this virtual component were uninstalled as well.

Task-number: QTIFW-3798
Change-Id: Ifd90d22651a97cfcf1e73f9b9fd74b85bb49f23b
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>