Fix calculation of installation size
Size calculation did not occur when user selected 'Select All/None/Reset'
in the componentselection page. Fixed so that when ever model changes,
the size is recalculated.

Also fixed the size to contain the size of maintenancetool when the
maintenancetool is not extracted from repository. If the maintenance
tool is fetched from repository, then the size is already calculated
when we calculate the size of extracted components.

Added also the size of local repository when needed. The size of the
local repository was not calculated correctly, the size is determined by
the components compressed size and not by application size.

Task-number: QTIFW-3757
Change-Id: I383d39e89b98ea2e62def1f7db19b8d2d8a47790
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>