Add operation for stopping processes
Introduce a new StopProcessesForUpdates operation. This operation
will be automatically invoked for components requesting processes
to be stopped via the component::addStopProcessForUpdateRequest()
script method, before installing, updating, or uninstalling the
component.

The operation has currently only effect on Windows - on Linux and
macOS it always returns true. The operation attempts to stop the
requested processes via the WM_CLOSE signal, or if a timeout is
reached, attempt to terminate it forcefully.

The component::addStopProcessForUpdateRequest() retains the old
behavior on Linux and macOS, that adds the FakeStopProcessesForUpdates
operation for the component, to support blocking the installation
action until requested processes are stopped manually by user.

Also fix handling of return value for the graceful process shut down
in sysinfo_win.cpp.

Task-number: QTIFW-3745
Task-number: QLS-1799
Change-Id: I2c6ed887fd9657d9560d6a0f945c654883ab2f0c
Reviewed-by: Katja Marttila <katja.marttila@qt.io>