Add possibility to style LabelWithPixmap widget
To use stylesheet in custom widget, paintEvent() needs override. To have
a QWidget size border set from stylesheet, sizeHint() needs to calculate
and return the needed width to have the correct border side. Normally
this would be done by adding space to layout after LabelWithPixmap, but
as the widget has a QLabel with wordwrap, the wordwrap causes the QLabel
to wrap the text earlier than needed.

Task-number: QTIFW-3740
Change-Id: Ie971bd7ccf46e81dc6d34a55b903f64093181edf
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>