Fix installation space calculation in macOS
macOS installer binary does not have a magic marker, so the
maintenance tool size calculation failed. Fixed so that if the
repository does not provide the maintenance tool, the current
binary size is used for maintenance tool size calculation.

Also reducing the verbosity of maintenance tool calculation error
message, as the fail is not significant for user since we have
a valid way of recovering from issue.

Change-Id: Ia367e4c29b5e4fd29a0e8e464c9d2c5c7b270d3f
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>