Remove empty folders when uninstalling
If an extract operation is defined in component's xml file, and the
extract path contains subfolders, installer creates those subfolders and
reports about created folders to operation. However, if several
components extracts to the same subfolder, the folder is already created
and no report is sent to operation.
Fixed so that after the extract operation is performed,  the directory
structure is saved for undo operation along with file list.

Task-number: QTIFW-3656
Change-Id: I9a9f462cb3fd384733d2b5c4d5193ddbeeaa643e
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>