3rdparty: remove bundled LZMA SDK
The bundled library has not been updated in a long time, and has been
marked deprecated for use with the IFW for several releases. Remove
the obsolete library and all references to it in the repository.

Users that have been still using theLZMA SDK archive backend in
self-built IFW should migrate to using libarchive, which is now enabled
by default.

Task-number: QTIFW-3819
Change-Id: I45298de282c2b0b65e8f343d90aa75495e55947a
Reviewed-by: Katja Marttila <katja.marttila@qt.io>