Fix repository action handling
We should not rely that the repository format is the same when we
compare repositories when adding, replacing or removing repositories
using RepositoryUpdate. This bug occured when building IFW with Qt 6.9.

Also, when using Qt6.7, QUrl does not resolve relative paths correctly
if the repository path contains duplicate trailing slashes.

Task-number: QTIFW-3730
Change-Id: Id8c6dd3702e6af484c81c817dbf7f4e7bb973b63
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>