Fix variable replace
Variables can be defined inside '@' symbols, and are resolved by
runtime, like @HomeDir@. This enables the usage of the correct variable
value before its final set.
If the variable inside '@' was not an installer value, installer
replaced the variable with empty value. As the double '@' can be also
part of the value itself, installer should not replace it with empty
value if no real variable is found for it.

Task-number: QTIFW-2945
Change-Id: I83ca40db1e03a12ba51535766b054765071d09ad
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>