Fix script usage for local installed packages
If local installed package tree name differs from tree name in remote,
the script is not found for the component nor evaluated. In situation
where is no tree names, local package is replaced with remote package,
with minor updates to the remote package, so that all component values
(for example new version) remains. As local package tree name is
overriding the remote value, local package is used instead of remote. As
the local package still needs the script information, release date
value, and new version, the local package is updated using the remote
values.

Task-number: QTIFW-3884
Change-Id: I1474f9cb8ba44639e44bedaacfb4c1ffd302cfd8
Reviewed-by: Katja Marttila <katja.marttila@qt.io>