Fix download error when network is unstable
If network is unstable, some packages might not be downloaded.
Installer will try to download missed packages again 3 times.
When retrying download, all necessary values were not cleared,
causing installer to download all files again and again, and
resulting to a failed installation and 'Cannot verify Hash' error.
Also, as we refetch the already fetched packages when hash verification
occurs, the downloaded amout exceeds the total amount. When this happens
we inform the user with a common status message instead of starting the
progress calculation from beginning.

Task-number: QTIFW-3882
Change-Id: I5d0b6a949e5c921d4f4fddcf5ae0c4b9085bedb4
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>