Fix mac maintenance tool update
macOS does not allow maintenance tool update with ditto when the
maintenance tool process is running. Created a new executable which
handles the maintenance tool update. The executable,
maintenanceToolUpdater, is launched after new maintenancetool is
available. The executable waits for the running maintenance tool to
finish, then deletes it, and updates the maintenance tool. macOS is
picky about removing app bundle, expecially if the tool is installed to
privileged folder. Removing immutable flags and extended attributes from
bundles which are going to be deleted allows Qt to recursively remove
app bundle.

MaintenanceToolUpdater should be injected to every repository build
which includes update to maintenance tool. It can be found in bin
-folder next to other executables.

Task-number: QTIFW-3759
Change-Id: I54c45e370e06f543eef3c86874c2695e25d88193
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>